LayoutTests/fast/mediastream/captureStream/canvas3d-expected.txt

CONSOLE MESSAGE: Turning drawing buffer preservation for the WebGL canvas being captured



PASS captureStream with 3d context drawing - not buffered initially
PASS captureStream with 3d context drawing - buffered initially
